Microsoft git installer options (#116034)
* microsoft/git: add custom installer arguments

These arguments are intended to tell the installer to be as quiet as
possible.

* /SUPPRESSMSGBOXES indicates that the installer should not create
  pop-up windows when an issue arises. Instead, the installer will fail.
  This will resolve https://github.com/microsoft/git/issues/595.

* "/enable-component windowsterminal" was added to the Git/Git manifest
  in #110011 to resolve #109915, which was creating some installation
  problems on Windows 10.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>

* microsoft/git: match some format for Git/Git's manifest

The Git/Git manifest has some additional options that were created
directly by the Winget team, as opposed to microsoft/git which was
created by the Git team. Since the Winget team is likely better at this
sort of thing, adopt some of the choices made over there.

These are mostly innocuous, though there is one interesting difference
that still remains: the Git/Git manifest has two installers with
different 'Scope' values ('machine' and 'user') while microsoft/git is
only available via 'machine' scope. This is purposefully left alone
since there is no difference in how the installers are specified and we
are not sure we want to support user-only installs of microsoft/git.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>

---------

Signed-off-by: Derrick Stolee <derrickstolee@github.com>